Refactor: 회원 탈퇴 로직 수정 (#139)
* Feat: 예외 코드 추가

* Feat: 파티 상태 변동에 따른 채팅방 관련 처리 추가

* Refactor: 파티 관련 코드에 facade 패턴 적용

* Refactor: 파티 facade 적용

* Refactor: 예약 관련 코드에 facade 패턴 적용

* Refactor: 예약 facade 적용

* Refactor: 유저 관련 코드에 facade 패턴 적용

* Refactor: 회원 탈퇴를 위한 쿼리 구현

* Refactor: 유저 facade 적용

* Refactor: oauth 회원가입 접근 권한 설정

* Test: 파티 관련 테스트에 facade 패턴 적용

* Test: 유저 관련 테스트에 facade 패턴 적용

* Test: 예약 관련 테스트에 facade 패턴 적용

* Feat: 예약 취소 api의 User 접근 권한 추가

* Refactor: 쿼리 정상화

* Refactor: 중복 코드 제거 및 주석 추가
