+++ matjalalzz/src/main/java/shop/matjalalzz/global/config/SecurityConfig.java
@@ -86,6 +86,7 @@ public class SecurityConfig {
                     .requestMatchers(HttpMethod.GET,"/owner/shops/{shopId}").hasAnyRole("ADMIN", "OWNER", "USER")
                     .requestMatchers(HttpMethod.PUT,"/owner/shops/{shopId}").hasAnyRole("ADMIN", "OWNER", "USER")
                     .requestMatchers(HttpMethod.GET, "/reservations").hasRole("OWNER")
+                    .requestMatchers(HttpMethod.PATCH, "/reservations/{reservationId}/cancel").hasRole("USER")
                     .requestMatchers(HttpMethod.PATCH, "/reservations/{reservationId}/**")
                     .hasRole("OWNER")
 
@@ -100,6 +101,7 @@ public class SecurityConfig {
                     .requestMatchers(HttpMethod.GET, "/users/set-cookie").permitAll()
 
                     .requestMatchers("/users/delete", "/users/logout").authenticated()
+                    .requestMatchers(HttpMethod.POST, "/users/signup/oauth").hasRole("TMP")
 
                     .anyRequest().hasAnyRole("USER", "ADMIN", "OWNER")
                 //나머지 요청은 USER 또는 ADMiN 권한을 가져야 접근 가능
