+++ matjalalzz/src/main/java/shop/matjalalzz/global/security/jwt/api/TokenController.java
@@ -1,23 +1,18 @@
 package shop.matjalalzz.global.security.jwt.api;
 
-import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletResponse;
-import java.net.URI;
 import java.util.HashMap;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import shop.matjalalzz.global.common.BaseResponse;
@@ -33,9 +28,6 @@ import shop.matjalalzz.global.security.jwt.dto.AccessTokenResponseDto;
 public class TokenController {
 
     private final TokenService tokenService;
-
-    @Value("${custom.jwt.redirect-login-success}")
-    private String redirectSuccess;
 
     @Operation(
         summary = "액세스 토큰 재발급",
