+++ matjalalzz/src/main/java/shop/matjalalzz/global/security/jwt/app/TokenService.java
@@ -1,9 +1,9 @@
 package shop.matjalalzz.global.security.jwt.app;
 
 import jakarta.servlet.http.HttpServletResponse;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shop.matjalalzz.global.exception.BusinessException;
@@ -11,7 +11,6 @@ import shop.matjalalzz.global.exception.domain.ErrorCode;
 import shop.matjalalzz.global.security.jwt.dao.RefreshTokenRepository;
 import shop.matjalalzz.global.security.jwt.dto.AccessTokenResponseDto;
 import shop.matjalalzz.global.security.jwt.dto.LoginTokenResponseDto;
-import shop.matjalalzz.global.security.jwt.dto.TokenBodyDto;
 import shop.matjalalzz.global.security.jwt.entity.RefreshToken;
 import shop.matjalalzz.global.security.jwt.mapper.TokenMapper;
 import shop.matjalalzz.global.util.CookieUtils;
@@ -50,6 +49,16 @@ public class TokenService {
         return TokenMapper.toLoginTokenResponseDto(accessToken, refreshToken.getRefreshToken());
     }
 
+    @Transactional
+    public RefreshToken saveRefreshToken(RefreshToken refreshToken) {
+        return refreshTokenRepository.save(refreshToken);
+    }
+
+    @Transactional
+    public void deleteRefreshToken(RefreshToken refreshToken) {
+        refreshTokenRepository.delete(refreshToken);
+    }
+
     @Transactional(readOnly = true)
     public AccessTokenResponseDto refreshAccessToken(String refreshToken) {
         if (!tokenProvider.validate(refreshToken)) {
@@ -65,6 +74,11 @@ public class TokenService {
         String newAccessToken = tokenProvider.issueAccessToken(user.getId(), user.getRole(),
             user.getEmail());
         return new AccessTokenResponseDto(newAccessToken);
+    }
+
+    @Transactional(readOnly = true)
+    public Optional<RefreshToken> findRefreshToken(User user) {
+        return refreshTokenRepository.findByUser(user);
     }
 
     @Transactional
