+++ matjalalzz/src/main/java/shop/matjalalzz/image/app/ImageService.java
@@ -1,0 +1,24 @@
+package shop.matjalalzz.image.app;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import shop.matjalalzz.image.dao.ImageRepository;
+
+@Service
+@RequiredArgsConstructor
+public class ImageService {
+
+    @Value("${aws.credentials.AWS_BASE_URL}")
+    private String BASE_URL;
+
+    private final ImageRepository imageRepository;
+
+    @Transactional(readOnly = true)
+    public String getShopThumbnail(Long shopId) {
+        return imageRepository.findByShopIdAndImageIndex(shopId, 0)
+            .map(image -> BASE_URL + image.getS3Key())
+            .orElse(null);
+    }
+}
