+++ matjalalzz/src/main/java/shop/matjalalzz/party/app/PartyStatusService.java
@@ -4,10 +4,12 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import shop.matjalalzz.chat.app.PartyChatService;
 import shop.matjalalzz.global.exception.BusinessException;
 import shop.matjalalzz.global.exception.domain.ErrorCode;
 import shop.matjalalzz.party.dao.PartyRepository;
 import shop.matjalalzz.party.entity.Party;
+import shop.matjalalzz.user.entity.User;
 
 @Service
 @RequiredArgsConstructor
@@ -15,6 +17,8 @@ import shop.matjalalzz.party.entity.Party;
 public class PartyStatusService {
 
     private final PartyRepository partyRepository;
+    private final PartyService partyService;
+    private final PartyChatService partyChatService;
 
     @Transactional
     public void updateStatusByDeadline(Long partyId) {
@@ -26,11 +30,16 @@ public class PartyStatusService {
             return;
         }
 
+        User host = partyService.findPartyHost(party);
+
         if (party.getCurrentCount() >= party.getMinCount()) {
             party.complete();
+            partyChatService.noticePaymentRequest(host, party);
         } else {
             party.deleteParty();
+            partyChatService.noticePartyDeleted(host, party);
         }
+
         log.info("파티 상태 변경 완료");
     }
