+++ matjalalzz/src/main/java/shop/matjalalzz/reservation/app/ReservationService.java
@@ -1,38 +1,21 @@
 package shop.matjalalzz.reservation.app;
 
-import static shop.matjalalzz.global.exception.domain.ErrorCode.ALREADY_PROCESSED;
-import static shop.matjalalzz.global.exception.domain.ErrorCode.DATA_NOT_FOUND;
-import static shop.matjalalzz.global.exception.domain.ErrorCode.FORBIDDEN_ACCESS;
-import static shop.matjalalzz.global.exception.domain.ErrorCode.SHOP_NOT_FOUND;
-
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shop.matjalalzz.global.exception.BusinessException;
 import shop.matjalalzz.global.exception.domain.ErrorCode;
-import shop.matjalalzz.party.app.PartyService;
 import shop.matjalalzz.party.entity.Party;
 import shop.matjalalzz.reservation.dao.ReservationRepository;
-import shop.matjalalzz.reservation.dto.CreateReservationRequest;
-import shop.matjalalzz.reservation.dto.CreateReservationResponse;
-import shop.matjalalzz.reservation.dto.MyReservationPageResponse;
 import shop.matjalalzz.reservation.dto.MyReservationResponse;
-import shop.matjalalzz.reservation.dto.ReservationListResponse;
-import shop.matjalalzz.reservation.dto.ReservationListResponse.ReservationContent;
 import shop.matjalalzz.reservation.entity.Reservation;
 import shop.matjalalzz.reservation.entity.ReservationStatus;
 import shop.matjalalzz.reservation.mapper.ReservationMapper;
-import shop.matjalalzz.shop.app.ShopService;
-import shop.matjalalzz.shop.entity.Shop;
-import shop.matjalalzz.user.app.UserService;
 import shop.matjalalzz.user.entity.User;
 
 @Slf4j
@@ -42,222 +25,67 @@ public class ReservationService {
 
     private final ReservationRepository reservationRepository;
 
-    private final ShopService shopService;
-    private final UserService userService;
-    private final PartyService partyService;
+    @Transactional
+    public void createPartyReservation(Party party, User host) {
+        Reservation reservation = ReservationMapper.toEntity(party, host);
+        reservationRepository.save(reservation);
+    }
+
+    @Transactional
+    public Reservation saveReservation(Reservation reservation) {
+        return reservationRepository.save(reservation);
+    }
+
+    @Transactional
+    public void refundPartyReservationFee(Party party) {
+        int fee = party.getShop().getReservationFee();
+        reservationRepository.refundPartyReservationFee(party.getId(), fee);
+    }
+
+    @Transactional
+    public void settleReservationFee(Long shopId, int reservationFee) {
+        reservationRepository.settleReservationFee(shopId, reservationFee);
+    }
 
     @Transactional(readOnly = true)
-    public ReservationListResponse getReservations(Long shopId, ReservationStatus status,
-        Long ownerId, Long cursor, int size) {
+    public Slice<MyReservationResponse> findByUserIdAndCursor(
+        Long userId, Long cursor, Pageable pageable) {
+        return reservationRepository.findByUserIdAndCursor(userId, cursor, pageable);
+    }
 
-        Pageable pageable = PageRequest.of(0, size, Sort.by(Direction.DESC, "id"));
-
-        if (shopId != null) {
-            Shop shop = shopService.shopFind(shopId); // 검증용
-
-            if (!shop.getUser().getId().equals(ownerId)) {
-                throw new BusinessException(FORBIDDEN_ACCESS);
-            }
-
-            Slice<Reservation> slice = reservationRepository.findByShopIdWithFilterAndCursor(
-                shopId, status, cursor, pageable
-            );
-
-            return reservationResponse(slice);
-        }
-
-        List<Shop> shops = shopService.findByOwnerId(ownerId);
-        if (shops == null) {
-            throw new BusinessException(SHOP_NOT_FOUND);
-        }
-
-        List<Long> shopIds = shops.stream()
-            .map(Shop::getId)
-            .toList();
-
-        Slice<Reservation> slice = reservationRepository.findByShopIdsWithFilterAndCursor(
+    @Transactional(readOnly = true)
+    public Slice<Reservation> findByShopIdsWithFilterAndCursor(
+        List<Long> shopIds, ReservationStatus status, Long cursor, Pageable pageable) {
+        return reservationRepository.findByShopIdsWithFilterAndCursor(
             shopIds, status, cursor, pageable
         );
-
-        return reservationResponse(slice);
-    }
-
-    private ReservationListResponse reservationResponse(Slice<Reservation> slice) {
-        List<Reservation> reservations = slice.getContent();
-
-        Long nextCursor = slice.hasNext()
-            ? reservations.getLast().getId()
-            : null;
-
-        List<ReservationContent> content =
-            ReservationMapper.toReservationContent(reservations);
-
-        return ReservationMapper.toReservationListResponse(content, nextCursor);
     }
 
     @Transactional(readOnly = true)
-    public MyReservationPageResponse findMyReservationPage(Long userId, Long cursor, int size) {
-        Slice<MyReservationResponse> reservations = reservationRepository.findByUserIdAndCursor(
-            userId, cursor,
-            PageRequest.of(0, size));
-
-        Long nextCursor = null;
-        if (reservations.hasNext()) {
-            nextCursor = reservations.getContent().getLast().reservationId();
-        }
-
-        return ReservationMapper.toMyReservationPageResponse(nextCursor, reservations);
-    }
-
-    @Transactional
-    public CreateReservationResponse createReservation(Long userId, Long shopId,
-        CreateReservationRequest request) {
-
-        Shop reservationShop = shopService.shopFind(shopId);
-        User reservationUser = userService.getUserById(userId);
-        int reservationFee = reservationShop.getReservationFee() * request.headCount();
-
-        if (reservationUser.getPoint() < reservationFee) {
-            throw new BusinessException(ErrorCode.LACK_OF_BALANCE);
-        }
-
-        LocalDateTime reservedAt = LocalDateTime.parse(request.date() + "T" + request.time());
-
-        Reservation reservation = ReservationMapper.toEntity(
-            request,
-            reservedAt,
-            reservationShop,
-            reservationUser
+    public Slice<Reservation> findByShopIdWithFilterAndCursor(
+        Long shopId, ReservationStatus status, Long cursor, Pageable pageable) {
+        return reservationRepository.findByShopIdWithFilterAndCursor(
+            shopId, status, cursor, pageable
         );
-
-        Reservation savedReservation = reservationRepository.save(reservation);
-
-        reservationUser.decreasePoint(reservationFee);
-
-        return ReservationMapper.toCreateReservationResponse(savedReservation);
     }
 
-    @Transactional
-    public void confirmReservation(Long reservationId, Long ownerId) {
-        Reservation reservation = validateOwnerPermissionAndPending(reservationId, ownerId);
+    @Transactional(readOnly = true)
+    public List<Reservation> findAllMyReservationByUserIdForWithdraw(Long userId) {
+        LocalDateTime threshold = LocalDateTime.now().plusDays(1);
 
-        reservation.changeStatus(ReservationStatus.CONFIRMED);
-    }
-
-    @Transactional
-    public void refuseReservation(Long reservationId, Long ownerId) {
-        Reservation reservation = validateOwnerPermissionAndPending(reservationId, ownerId);
-
-        refuseReservation(reservation);
-    }
-
-    @Transactional
-    public void cancelReservation(Long reservationId, Long userId) {
-        Reservation reservation = getReservationById(reservationId);
-        User user = userService.getUserById(userId);
-
-        // 예약한 본인인지 확인
-        if (!reservation.getUser().getId().equals(userId)) {
-            throw new BusinessException(FORBIDDEN_ACCESS);
-        }
-
-        Party party = reservation.getParty();
-        if (party != null) {
-            partyService.breakParty(party);
-            return;
-        }
-
-        ReservationStatus status = reservation.getStatus();
-
-        // 예약 상태 확인
-        if (status != ReservationStatus.CONFIRMED && status != ReservationStatus.PENDING) {
-            throw new BusinessException(ALREADY_PROCESSED);
-        }
-
-        // 예약이 승인됐고, 예약일로부터 하루도 안남았으면, 취소 불가
-        if (status == ReservationStatus.CONFIRMED &&
-            reservation.getReservedAt().isBefore(LocalDateTime.now().plusDays(1))) {
-            throw new BusinessException(ErrorCode.CANNOT_CANCEL_D_DAY);
-        }
-
-        user.increasePoint(reservation.getReservationFee());
-        reservation.changeStatus(ReservationStatus.CANCELLED);
-    }
-
-    private Reservation validateOwnerPermissionAndPending(Long reservationId,
-        Long ownerId) {
-        Reservation reservation = getReservationById(reservationId);
-
-        if (reservation.getStatus() != ReservationStatus.PENDING) {
-            throw new BusinessException(ALREADY_PROCESSED);
-        }
-
-        if (!reservation.getShop().getUser().getId().equals(ownerId)) {
-            throw new BusinessException(FORBIDDEN_ACCESS);
-        }
-
-        return reservation;
+        // 회원 단독으로 진행한 예약 중 대기 상태이거나 예약일로부터 하루 이상 남은 예약 조회
+        return reservationRepository.findAllMyReservationByUserIdForWithdraw(userId, threshold);
     }
 
     @Transactional(readOnly = true)
     public Reservation getReservationById(Long reservationId) {
         return reservationRepository.findById(reservationId)
-            .orElseThrow(() -> new BusinessException(DATA_NOT_FOUND));
+            .orElseThrow(() -> new BusinessException(ErrorCode.DATA_NOT_FOUND));
     }
 
-    @Transactional
-    public int terminateExpiredReservations() {
-        LocalDateTime threshold = LocalDateTime.now().minusDays(1);
-
-        List<Reservation> toTerminate = reservationRepository
-            .findAllByStatusAndReservedAtBefore(
-                ReservationStatus.CONFIRMED,
-                threshold
-            );
-
-        for (Reservation r : toTerminate) {
-            r.changeStatus(ReservationStatus.TERMINATED);
-
-            Party party = r.getParty();
-
-            if (party != null) {
-                party.terminate();
-            }
-
-            // 예약금 정산
-            reservationRepository.settleReservationFee(r.getShop().getId(), r.getReservationFee());
-        }
-
-        return toTerminate.size();
-    }
-
-    @Transactional
-    public int refuseExpiredPendingReservations() {
-        LocalDateTime threshold = LocalDateTime.now().minusHours(1);
-
-        List<Reservation> toRefuse = reservationRepository
-            .findAllByStatusAndReservedAtBefore(
-                ReservationStatus.PENDING,
-                threshold
-            );
-
-        for (Reservation r : toRefuse) {
-            refuseReservation(r);
-        }
-
-        return toRefuse.size();
-    }
-
-    private void refuseReservation(Reservation reservation) {
-        Party party = reservation.getParty();
-
-        if (party != null) {
-            partyService.breakParty(party);
-        } else {
-            reservation.getUser().increasePoint(reservation.getReservationFee());
-        }
-
-        reservation.changeStatus(ReservationStatus.REFUSED);
+    @Transactional(readOnly = true)
+    public List<Reservation> findAllByStatusAndReservedAtBefore(ReservationStatus status, LocalDateTime threshold) {
+        return reservationRepository.findAllByStatusAndReservedAtBefore(status, threshold);
     }
 }
